Action to assign PR to its owner
Makes available a GitHub Action which retrieves the creator of a Pull
Request and assigns it to him. GitHub API will silently fail to assign
the pull request if the user does not have `push` privileges to the
repository, improvements can be made in the future trough a *personal
token* granted it has access to fetch information about
repository/organization collaborators - which may prove too much to this
action.

Tests are marked as TODOs, don't know TypeScript enough and I am
unwilling to spend too much time on just this right now.

Releases to Github Package Registry are managed through an action as
well, triggered by the creation of a release on GitHub. An improvement
could be made to ensure the version on `package.json` corresponds to the
tag and release being created.