Bumps patch version on `package.json`
I've made a mess testing this, by successfully creating a package with
the 1.0.0 version and failing to unpublish with with the existing
workflow (although it executes successfully).

This version bump ensures the package exists with the right version of
the code when the tag and release is created on GitHub.